[FIX] Fix using the appropriate types and returns

+++ median_calculation/median_calculation--0.0.1.sql
@@ -1,3 +1,3 @@
-CREATE OR REPLACE FUNCTION median_calculation(text, text) RETURNS numeric 
+CREATE OR REPLACE FUNCTION median_calculation(text, text) RETURNS float 
     AS 'MODULE_PATHNAME', 'median_calculation'
     LANGUAGE C STRICT;
